LayoutTests/inspector/dom-debugger/event-listener-breakpoints-builtin-exact-case-insensitive-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for Event Listener breakpoints.


== Running test suite: DOMDebugger.Event.Listener
-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.AddBreakpoint
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...
Firing "click" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_click
1: [F] trigger_click
2: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.AddDisabledBreakpoint
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...
Disabling "listener:cLiCk" Event Breakpoint...
Firing "click" on body...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.RemoveBreakpoint
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...
Removing "listener:cLiCk" Event Breakpoint...
Firing "click" on body...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.RemoveDisabledBreakpoint
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...
Disabling "listener:cLiCk" Event Breakpoint...
Removing "listener:cLiCk" Event Breakpoint...
Firing "click" on body...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.Options.Condition
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.Options.Condition.ConsoleCommandLineAPI
Adding saved console value 'false'...
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...

Setting condition to saved console value...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Adding saved console value 'true'...
Setting condition to saved console value...

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.Options.IgnoreCount
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should not pause.

Triggering breakpoint...
PASS: Should pause.

Triggering breakpoint...
PASS: Should pause.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.Options.Action.Log
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...

Adding log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.Options.Actions.Evaluate
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...

Adding evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: DOMDebugger.Event.Listener.Builtin.Exact.CaseInsensitive.Options.Actions.Evaluate.ConsoleCommandLineAPI
Adding saved console value '1'...
Creating "cLiCk" Event Breakpoint...
Adding "listener:cLiCk" Event Breakpoint...

Adding evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '2'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should pause.

Adding saved console value '3'...
Editing evaluate action using saved console value...
Enabling auto-continue...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Adding saved console value '4'...
Editing evaluate action using saved console value...

Triggering breakpoint...
PASS: Should execute breakpoint action.
PASS: Should not pause.
